Modularize the code
