+++ index.js
@@ -1,1 +1,2 @@
-export {default as jch, jab, interpolateJab} from "./src/cam02";
+export {default as jch, jab} from "./src/cam02";
+export {default as interpolateJab} from "./src/interpolateJab";
